+++ CHANGELOG.md
@@ -3,6 +3,8 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 
+* Allow passing `ignoreChanges` into `Subnet`s created as part of an `awsx.ec2.Vpc`.
+
 ## 0.18.10 (2019-08-21)
 
 * Updated `@pulumi/awsx` to use the latest version of `@pulumi/docker`.